backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain
langchain-openai
langchain-chroma
langgraph
chromadb
python-multipart
anthropic
httpx
langfuse>=2.0.0,<3.0.0 
google-generativeai
tavily-python